Rename ZsrClient to OAClient in C++.
