+++ libs/zswagcl/include/zswagcl/private/openapi-client.hpp
