+++ libs/zswagcl/src/openapi-parser.cpp
@@ -1,4 +1,4 @@
-#include "openapi-parser.hpp"
+#include "private/openapi-parser.hpp"
 
 #include "httpcl/http-settings.hpp"
 #include "httpcl/uri.hpp"
